modules/StoreTemperature/requirements.txt
azure-iot-device~=2.1.4
azure-storage-blob
azure-core